add async actions to redux

+++ src/components/cart/cart.component.tsx
@@ -28,7 +28,7 @@ import {
 @containerComponent
 export class Cart extends Component<Props> {
   handdleClick = data => () => {
-    this.props.deleteAction({ id: data.id });
+    if (this.props.deleteAction) this.props.deleteAction({ id: data.id });
   };
 
   render() {
@@ -37,21 +37,23 @@ export class Cart extends Component<Props> {
         <Image src={cart} />
         <CartArea className="cart">
           <CartBox>
-            {this.props.cart.map(i => (
-              <BoxMiniCart key={i.id}>
-                <MiniImg src={`http://localhost:3000${i.image}`} />
-                <MiniDataArea>
-                  <MiniTxt>{i.name}</MiniTxt>
-                  <MiniTxt bold>
-                    {i.price}
-                    {i.quantity > 1 ? ` x${i.quantity}` : null}
-                  </MiniTxt>
-                </MiniDataArea>
-                <MiniCloseArea>
-                  <Link onClick={this.handdleClick(i)}>X</Link>
-                </MiniCloseArea>
-              </BoxMiniCart>
-            ))}
+            {this.props.cart
+              ? this.props.cart.map(i => (
+                  <BoxMiniCart key={i.id}>
+                    <MiniImg src={`http://localhost:3000${i.image}`} />
+                    <MiniDataArea>
+                      <MiniTxt>{i.name}</MiniTxt>
+                      <MiniTxt bold>
+                        {i.price}
+                        {i.quantity > 1 ? ` x${i.quantity}` : null}
+                      </MiniTxt>
+                    </MiniDataArea>
+                    <MiniCloseArea>
+                      <Link onClick={this.handdleClick(i)}>X</Link>
+                    </MiniCloseArea>
+                  </BoxMiniCart>
+                ))
+              : null}
           </CartBox>
         </CartArea>
       </Container>
